make `make test[cluster]` write clean check.*.set
formerly `make test[cluster]` wrote just copied the settings file of
the last run instance, now it loads and writes it out with a EXECUTABLE.
This was a problem because for some instances special additional settings
files were loaded

+++ check/evalcheck_cluster.sh
@@ -149,7 +149,10 @@ do
         FILE=$i.set
         if test -e $FILE
         then
-            cp $FILE $SETFILE
+            if ! test -e $SETFILE
+            then
+               cp $FILE $SETFILE
+            fi
             if test "$REMOVE" = "1"
             then
                 rm -f $FILE
